3.30_country_agent/agents/pyproject.toml
[project]
name = "adk-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["google-adk"]
